+++ CanInterfaceImplementations/systec/STCanScan.h
@@ -70,6 +70,10 @@ public:
 	virtual CanModule::ReconnectAutoCondition getReconnectCondition() { return m_reconnectCondition; };
 	virtual CanModule::ReconnectAction getReconnectAction() { return m_reconnectAction; };
 
+	virtual void stopBus(){
+		MLOGST(Log::TRC) << __FUNCTION__ << " (doing nothing)";
+	};
+
 private:
 
 	int m_moduleNumber;